Re-enable remote attestation at the server
Since I've managed to successfully install Omeg's `pod_client` lib,
and generate a valid quote, the IAS attestation now works fine within
the web service, so I'm gonna go ahead and re-enable it then.

Also, while exploring this, I've found a small bug with `rust-sgx-util`
in the way we were calling `ias_verify_quote` C-fn, and I've fixed it
now.

It is now possible to send the `pod_client` generated quote using
the `test_client` to the server and properly register the user.
However, the authentication bit is currently not functional since I
still need to hook up `pod_client` lib into the `test_client` to
be able to spin up the enclave and sign the challenge.