Don't export public key to file
Given that the public key is already embedded in the exported quote,
I don't think we need to have separate functionality for exporting
our public key to file upon enclave initialisation. This commit
effectively removes this functionality in favour of dealing with
the quote directly for public key extraction (which is already being
done on the web server side anyhow BTW).