Add simple check for config.toml for pod-app
Instead of only hard-coding the values for the path to the sealed private
key and the enclave, we now search for `config.toml` in `XDG_CONFIG_HOME`.
If it is not found, then we use the hard-coded paths to the private key
and the enclave.

In general, the latter is expected to be mainly used when building the
app from source for development and debugging. The former should be
utilised when installing the app using the provided installation
scripts.