Config and installation script for pod-app (#21)
* Add simple check for config.toml for pod-app

Instead of only hard-coding the values for the path to the sealed private
key and the enclave, we now search for `config.toml` in `XDG_CONFIG_HOME`.
If it is not found, then we use the hard-coded paths to the private key
and the enclave.

In general, the latter is expected to be mainly used when building the
app from source for development and debugging. The former should be
utilised when installing the app using the provided installation
scripts.

* Add simple install script

The simple install script will run `cargo install` of the `pod-app`
which will place `pod-app` in local cargo registry. It will also
crate `$HOME/.local/share/pod-app` directory and copy
`pod_enclave.signed.so` into to. The sealed private key will also
by definition reside in that folder.

* Update pod-app/README.md