Add logging to file in pod-app
Logging `pod-app` is somewhat awkward since we use `stdin` and `stdout`
as comms channels to speak with the browser extension. For debugging
however it would good to have some info about what is going on
in the app. This commit enables logging to a file of the internals
of `pod-app`.